fix countgd
