+++ .github/workflows/ci_cd.yml
@@ -83,9 +83,6 @@ jobs:
     - name: Test with pytest
       run: |
         poetry run pytest -v tests/integ
-    - name: Test with pytest, dev env
-      run: |
-        LANDINGAI_API_KEY=$LANDINGAI_DEV_API_KEY LANDINGAI_URL=https://api.dev.landing.ai poetry run pytest -v tests/integration_dev
 
   release:
     name: Release
